add resolver for concept map title
