+++ client/src/graphql/mutations/updateInput.graphql
@@ -2,6 +2,6 @@ mutation updateInput($inputId: ID!, $data: UpdateInputInput!) {
   updateInput(inputId: $inputId, data: $data) {
     id
     script
-    conceptMapId
+    conceptMap
   }
 }
